pyproject.toml
[project]
name = "langgraphprac"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.7.0",
    "langchain>=1.0.5",
    "langchain-core>=1.0.4",
    "langchain-google-genai>=4.1.2",
    "langchain-google-vertexai>=3.0.3",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "msgpack>=1.1.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "deepagents>=0.3.5",
    "tavily-python>=0.7.17",
    "langchain-community>=0.4.1",
    "langchain-pymupdf4llm>=0.5.0",
    "pillow>=12.1.0",
    "rapidocr-onnxruntime>=1.4.4",
    "ipykernel>=7.1.0",
    "boto3>=1.42.28",
    "python-multipart>=0.0.21",
    "langchain-text-splitters>=1.1.0",
    "python-pptx>=1.0.2",
]
